Added the django-inspectional-registration module
We need people to create accounts on the site: but don't want just anyone
to be able to sign up - only those who we want. As a short term measure
we'll use this module.

But there are one or two issues:
1) It doesn't seem to use AUTH_PASSWORD_VALIDATORS when allowing
    users to set their passwords :(
2) It doesn't seem to email the admin when a user registers. Rather, the
    admins have to inspect the registration requests on a regular basis :(

If we can't resolve the issues we may have to go down some other path.