Added start of 'report references link'
It would be a good idea if the extra report on cloud active projects
(the cloud top twenty) was not hard coded into the page. Then other
pages would not need to display the extra html, and the reports
would be more compo
sable.

This is the first step in moving toward that more flexible future.