Extended list of cell names
NeCTAR cunningly removed their middle cell's and shortened the names of the child cells.
This meant that we were no longer picking up the instances that ran/run with with the
new cell names. So changed the cell names in the settings file, and parametrised the
queries to use the settings rather than to have the names hard coded.