Removed the quarter start dates and renamed end date column
The javascript is written to automatically filter on a column named 'date'. By
breaking that model we either need to add new code to handle it, or fit the
data to the existing code. It was decided that it was simpler to fit the data to
the existing code for the time being.